+++ js/jquery.terminal.js
@@ -41,7 +41,7 @@
  *
  * broken image by Sophia Bai from the Noun Project (CC-BY)
  *
- * Date: Wed, 08 Sep 2021 16:45:06 +0000
+ * Date: Wed, 08 Sep 2021 16:51:59 +0000
  */
 /* global define, Map */
 /* eslint-disable */
@@ -5033,7 +5033,7 @@
     // -------------------------------------------------------------------------
     $.terminal = {
         version: 'DEV',
-        date: 'Wed, 08 Sep 2021 16:45:06 +0000',
+        date: 'Wed, 08 Sep 2021 16:51:59 +0000',
         // colors from https://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',
@@ -9826,12 +9826,18 @@
                     var cmd_prompt = self.find('.cmd-prompt');
                     var cmd_outer = self.find('.cmd');
                     partial = self.find('.partial');
+                    var last_row;
                     if (partial.length === 0) {
                         cmd_prompt.css('margin-left', 0);
                         cmd_outer.css('top', 0);
                         command_line.__set_prompt_margin(0);
+                        last_row = self.find('.terminal-output div:last-child div:last-child');
+                        last_row.css({
+                            width: '100%',
+                            display: ''
+                        });
                     } else {
-                        var last_row = partial.children().last();
+                        last_row = partial.children().last();
                         // Remove width='100%' for two reasons:
                         // 1. so we can measure the width right here
                         // 2. so that the background of this last line of output